+++ src/components/KpiWidgetGroup.vue
@@ -19,31 +19,35 @@
         </pkt-button>
       </template>
     </div>
-    <div class="kpi-widget-group__kpis">
-      <router-link
-        v-for="kpi in kpis"
-        :key="kpi.id"
-        :to="{
-          name: 'ItemMeasurements',
-          params: { ...$route.params, kpiId: kpi.id },
-          query: { resultIndicatorPeriod: selectedPeriod?.key },
-        }"
-        class="kpi-widget-group__link"
-      >
-        <widget-kpi-card :kpi="kpi" :compact="compact" />
-      </router-link>
-    </div>
+    <draggable v-model="orderedKpis" :options="{ animation: 200 }">
+      <transition-group class="kpi-widget-group__kpis">
+        <router-link
+          v-for="kpi in orderedKpis"
+          :key="kpi.id"
+          :to="{
+            name: 'ItemMeasurements',
+            params: { ...$route.params, kpiId: kpi.id },
+            query: { resultIndicatorPeriod: selectedPeriod?.key },
+          }"
+          class="kpi-widget-group__link"
+        >
+          <widget-kpi-card :kpi="kpi" :compact="compact" />
+        </router-link>
+      </transition-group>
+    </draggable>
   </div>
 </template>
 
 <script>
+import draggable from 'vuedraggable';
 import html2canvas from 'html2canvas';
 import { mapState } from 'vuex';
 import { periodDates } from '@/util';
 import { PktButton } from '@oslokommune/punkt-vue2';
 import downloadFile from '@/util/downloadFile';
-import { formatKPIValue } from '@/util/kpiHelpers';
+import Kpi from '@/db/Kpi';
 import WidgetKpiCard from '@/components/widgets/WidgetKpiCard/WidgetKpiCard.vue';
+import { compareKPIs, formatKPIValue } from '@/util/kpiHelpers';
 
 export default {
   name: 'KpiWidgetGroup',
@@ -51,6 +55,7 @@ export default {
   components: {
     PktButton,
     WidgetKpiCard,
+    draggable,
   },
 
   props: {
@@ -91,6 +96,19 @@ export default {
         ...this.title.toLowerCase().split(' '),
         now.toISOString().slice(0, 10),
       ].join('-');
+    },
+
+    orderedKpis: {
+      get() {
+        return this.kpis.map((kpi) => kpi).sort(compareKPIs);
+      },
+      set(kpis) {
+        kpis.forEach((kpi, i) => {
+          if (kpi.order !== i) {
+            Kpi.update(kpi.id, { order: i });
+          }
+        });
+      },
     },
   },