+++ src/util/kpiHelpers.js
@@ -218,3 +218,35 @@ export function filterDuplicatedProgressValues(progressCollection) {
     return false;
   });
 }
+
+const _kpiOrder = ['ri', 'keyfig', 'plain'];
+
+/**
+ * Compare KPIs `a` and `b` for sorting.
+ *
+ * Return a negative number if `a` should come before `b`, otherwise return a
+ * positive number.
+ *
+ * KPIs are ordered on two levels, first by KPI type, then by their set order
+ * within that type. To retain backwards compatibility for KPIs that haven't
+ * gotten any order set yet, they're ordered by their name in that case.
+ */
+export function compareKPIs(a, b) {
+  // First sort by KPI type.
+  if (a.kpiType !== b.kpiType) {
+    return _kpiOrder.indexOf(a.kpiType) - _kpiOrder.indexOf(b.kpiType);
+  }
+  // Then sort by order only if both have one.
+  if ('order' in a && 'order' in b) {
+    return a.order - b.order;
+  }
+  // When only one of them has an order, sort that one first.
+  if ('order' in a) {
+    return -1;
+  }
+  if ('order' in b) {
+    return 1;
+  }
+  // Otherwise fall back to ordering by name.
+  return a.name.localeCompare(b.name);
+}