+++ CHANGELOG.md
@@ -29,6 +29,8 @@ All notable changes to this project will be documented in this file. The format 
 - The CSV export of measurements no longer includes the time part of the
   timestamps, as only one measurement can exist per day anyway.
 - Line breaks in measurement comments are now rendered.
+- Measurements can now be rearranged by drag and drop on the measurements
+  overview page.
 
 ### Removed